Code generation nuget
