+++ buildConf/Generic-CodeGen.props
@@ -1,0 +1,14 @@
+<Project>
+ <PropertyGroup>
+   <PackageTags>Netherum;Ethereum;Blockchain;Crypto 2.0</PackageTags>
+   <PackageIconUrl>https://raw.githubusercontent.com/Nethereum/Nethereum/master/logos/logo64x64.png</PackageIconUrl>
+   <PackageProjectUrl>http://www.nethereum.com</PackageProjectUrl>
+   <PackageLicenseUrl>https://raw.githubusercontent.com/Nethereum/Nethereum/master/LICENSE.md</PackageLicenseUrl>
+   <RepositoryType>git</RepositoryType>
+   <RepositoryUrl>https://github.com/Nethereum/Nethereum</RepositoryUrl>
+   <Description>Code generation of Nethereum integration classes using the Solidity Abi and Bin</Description>
+   <Copyright>Juan Blanco</Copyright>
+   <Version>1.0.1</Version>
+   <Authors>Juan Blanco, Dave Whiffin</Authors> 
+</PropertyGroup>
+</Project>
