+++ src/Nethereum.Generators.Net/ContractLibrary/GenerateClassesCommand.cs
@@ -2,25 +2,6 @@ using Nethereum.Generators.Core;
 
 namespace Nethereum.Generators.Net.ContractLibrary
 {
-
-    public class DefaultNamespaceService
-    {
-        public static string SetDefaultCqs(string contractName)
-        {
-            return contractName + "." + "CQS";
-        }
-
-        public static string SetDefaultDto(string contractName)
-        {
-            return contractName + "." + "DTOs";
-        }
-
-        public static string SetDefaultService(string contractName)
-        {
-            return contractName + "." + "Service";
-        }
-    }
-
     public class GenerateClassesCommand
     {
         public string ContractByteCode { get; set; }
