+++ src/nuget-push-generators.bat
@@ -1,0 +1,3 @@
+cd packages-generators
+FOR %%i IN (*.nupkg) DO dotnet nuget push %%i -s https://www.nuget.org/api/v2/package
+cd..
